condor: v3.4 → v4.0.0 helpers-native migration (plumbing-only)
Condor migrated from legacy runtime-1.0-era pattern (Python full-agency
scanner + openclaw cron + mcporter subprocess + internal create_position)
to helpers-native runtime-1.1.0 pattern. Plumbing-only — NO thesis
change. v3.4 hard gates (3TF alignment, MACRO_TREND_GATE, SM consensus
70%, OI/trader/XYZ filters), scoring tables, score-scaled sizing
(50%/70%/80% margin), 10x leverage cap, and DSL preset (Kodiak SOL
empirical 6-tier ladder) preserved verbatim.

Six-layer plumbing flip:
1. MCP transport: mcporter subprocess (2.5-5s cold-start) →
   SenpiClient.mcp_call() in-process HTTPS (~280ms).
2. Signal emit: scanner called create_position directly; producer
   now emits via push_signal() to runtime /signals. Runtime opens
   via LLM-gated condor_entry action with FEE_OPTIMIZED_LIMIT,
   passing marginUsd/leverage through from signal data for
   score-tier sizing.
3. Reentrancy: no v3 lockfile to replace; producer_daemon owns
   scanner_lock with stale-PID auto-recovery.
4. Scheduler: openclaw cron (3min) → producer_daemon(
   interval_seconds=180). Long-lived process.
5. Risk gates: Python MAX_POSITIONS + dynamic daily cap (1 healthy
   / 0 at -25% drawdown) + post-exit cooldown state files →
   declarative risk.guard_rails (max_entries_per_day=1,
   per_asset_cooldown_minutes=120, drawdown_halt_pct=25).
6. Exit fee: DSL exits MARKET (taker 0.045%) → FEE_OPTIMIZED_LIMIT
   (maker-first 0.015%, 60s ALO timeout, taker fallback). Entries
   keep ensure_execution_as_taker: false per v3.x patience rule.

Also fixed: condor_config.py had a copy-paste docstring header saying
"MANTIS Strategy" with mantis-strategy paths — corrected to Condor.

Files:
- condor/scripts/condor-producer.py (NEW; replaces condor-scanner.py)
- condor/scripts/condor_config.py (rewrite: drop mcporter subprocess,
  add SDK probe + lazy SenpiClient wrapper, fix MANTIS-mislabeled
  header, mcp_call routes through SenpiClient, mcporter_call alias
  preserved for compat)
- condor/runtime.yaml (add external_scanner condor_signals, LLM-gated
  condor_entry with score-tier marginUsd/leverage pass-through,
  risk.guard_rails, FEE_OPTIMIZED_LIMIT entries+exits; DSL preset
  preserved verbatim from v3.3+/v3.4 with weak_peak_cut DISABLED)
- condor/SKILL.md (v4.0.0 frontmatter + architecture; thesis preserved)
- condor/README.md (new helpers-native install — Step 0 openclaw.json,
  Step 1 install runtime skill, Step 2 pull files, Step 3 config.json,
  Step 4 env vars, Step 5 recreate runtime + launch daemon)
- condor/scripts/condor-scanner.py (deleted)
