+++ condor/scripts/condor_config.py
@@ -1,30 +1,71 @@
-"""MANTIS Strategy — Shared config, MCP helpers, state I/O.
-Self-contained — does not depend on wolf_config."""
+"""CONDOR v4.0.0 — Shared config + MCP shim + helpers wrapper.
+
+v4.0.0: senpi_runtime_helpers migration. mcporter_call now routes
+through SenpiClient.mcp_call() (direct HTTPS) instead of mcporter
+subprocess. _wrapper_client is exposed for the producer's signal
+push (cfg._wrapper_client.push_signal(...)).
+"""
 # Copyright 2026 Senpi (https://senpi.ai)
 # Licensed under MIT
 # Source: https://github.com/Senpi-ai/senpi-skills
 
+import functools
 import json
 import os
-import subprocess
 import sys
 import tempfile
 import time
-import glob
 from datetime import datetime, timezone
 from pathlib import Path
 
 WORKSPACE = os.environ.get("OPENCLAW_WORKSPACE", "/data/workspace")
-SKILL_DIR = Path(WORKSPACE) / "skills" / "mantis-strategy"
-CONFIG_PATH = SKILL_DIR / "config" / "mantis-config.json"
+SKILL_DIR = Path(WORKSPACE) / "skills" / "condor-strategy"
+CONFIG_PATH = SKILL_DIR / "config" / "condor-config.json"
 STATE_DIR = SKILL_DIR / "state"
-HISTORY_FILE = os.path.join(WORKSPACE, "mantis-emerging-history.json")
-COOLDOWN_FILE = STATE_DIR / "asset-cooldowns.json"
 
 STATE_DIR.mkdir(parents=True, exist_ok=True)
 
 
-# ─── Atomic Write ────────────────────────────────────────────
+# ─── senpi_runtime_helpers (lazy + auth-validated) ───
+# senpi_runtime_helpers ships inside the senpi-trading-runtime skill.
+# Global skills install under ~/.openclaw/skills/ on standard hosts
+# (e.g. /data/.openclaw/skills/ on Railway). Some setups install user
+# skills under ${OPENCLAW_WORKSPACE}/skills/. Probe both in order.
+_sdk_candidates = [
+    str(Path.home() / ".openclaw" / "skills" / "senpi-trading-runtime"),
+    str(Path(os.environ.get("OPENCLAW_WORKSPACE", "/data/workspace")) / "skills" / "senpi-trading-runtime"),
+]
+_sdk_path = next(
+    (p for p in _sdk_candidates if (Path(p) / "senpi_runtime_helpers").is_dir()),
+    _sdk_candidates[0],
+)
+if _sdk_path not in sys.path:
+    sys.path.insert(0, _sdk_path)
+from senpi_runtime_helpers import SenpiClient, log_event  # type: ignore  # noqa: E402
+
+
+@functools.lru_cache(maxsize=1)
+def _get_wrapper_client() -> SenpiClient:
+    if not os.environ.get("SENPI_AUTH_TOKEN", "").strip():
+        raise RuntimeError(
+            "SENPI_AUTH_TOKEN is not set. Condor's MCP calls and signal "
+            "POST both require it. Set it on the runtime host before "
+            "starting the producer daemon."
+        )
+    client = SenpiClient()
+    log_event("condor_wrapper_enabled", sdk_path=_sdk_path)
+    return client
+
+
+class _WrapperClientProxy:
+    def __getattr__(self, name: str):
+        return getattr(_get_wrapper_client(), name)
+
+
+_wrapper_client = _WrapperClientProxy()
+
+
+# --- Atomic Write ---
 
 def atomic_write(path, data):
     """Write JSON atomically via tmp file + os.replace."""
@@ -44,7 +85,7 @@ def atomic_write(path, data):
         raise
 
 
-# ─── Config ──────────────────────────────────────────────────
+# --- Config ---
 
 def load_config():
     if CONFIG_PATH.exists():
@@ -54,52 +95,34 @@ def load_config():
 
 
 def get_wallet_and_strategy():
-    wallet = os.environ.get("MANTIS_WALLET", "")
-    strategy_id = os.environ.get("MANTIS_STRATEGY_ID", "")
+    """Resolve (wallet, strategyId) — env var first, config.json second.
+
+    Env var resolution for wallet (first non-empty wins):
+      1. CONDOR_WALLET — fleet-standard <SKILL>_WALLET name (v2.0.9 rule)
+      2. cfg.load_config()["wallet"] — canonical source on disk
+    """
+    wallet = os.environ.get("CONDOR_WALLET", "").strip()
+    strategy_id = os.environ.get("CONDOR_STRATEGY_ID", "").strip()
     if not wallet or not strategy_id:
         config = load_config()
-        wallet = wallet or config.get("wallet", "")
-        strategy_id = strategy_id or config.get("strategyId", "")
+        wallet = wallet or config.get("wallet", "").strip()
+        strategy_id = strategy_id or config.get("strategyId", "").strip()
     return wallet, strategy_id
 
 
-# ─── State I/O ───────────────────────────────────────────────
-
-def load_state(filename="state.json"):
-    path = STATE_DIR / filename
-    if path.exists():
-        with open(path) as f:
-            return json.load(f)
-    return {}
-
-
-def save_state(data, filename="state.json"):
-    atomic_write(str(STATE_DIR / filename), data)
-
-
-# ─── Trade Counter ───────────────────────────────────────────
+# --- Trade counter (preserved for post-exit cooldown tracking only) ---
 
 def load_trade_counter():
     today = datetime.now(timezone.utc).strftime("%Y-%m-%d")
     path = STATE_DIR / "trade-counter.json"
-    default = {
-        "date": today, "entries": 0, "realizedPnl": 0,
-        "gate": "OPEN", "gateReason": None, "cooldownUntil": None,
-        "lastResults": [],
-        "stalkerResults": [],  # v1.2: track Stalker W/L for streak detection
-    }
+    default = {"date": today, "entries": 0, "last_entry_ts": 0}
     if path.exists():
         try:
             with open(path) as f:
                 tc = json.load(f)
             if tc.get("date") != today:
-                for k in ["entries", "realizedPnl"]:
-                    tc[k] = 0
                 tc["date"] = today
-                tc["gate"] = "OPEN"
-                tc["gateReason"] = None
-                tc["cooldownUntil"] = None
-                # NOTE: stalkerResults persists across days (streak spans sessions)
+                tc["entries"] = 0
             for k, v in default.items():
                 if k not in tc:
                     tc[k] = v
@@ -114,110 +137,35 @@ def save_trade_counter(tc):
     atomic_write(str(STATE_DIR / "trade-counter.json"), tc)
 
 
-def record_stalker_result(tc, is_win):
-    """v1.2: Track Stalker trade results for streak detection.
-    If a win, reset the streak. Keep last 10 results."""
-    results = tc.get("stalkerResults", [])
-    results.append("W" if is_win else "L")
-    tc["stalkerResults"] = results[-10:]
-    save_trade_counter(tc)
+# --- MCP Helper ---
 
+def mcp_call(tool, **params):
+    """Direct MCP call via SenpiClient (in-process HTTPS).
 
-# ─── Asset Cooldowns ─────────────────────────────────────────
-
-def load_cooldowns():
-    if COOLDOWN_FILE.exists():
-        try:
-            with open(COOLDOWN_FILE) as f:
-                return json.load(f)
-        except (json.JSONDecodeError, IOError):
-            pass
-    return {}
-
-
-def save_cooldowns(cooldowns):
-    atomic_write(str(COOLDOWN_FILE), cooldowns)
-
-
-def is_asset_cooled_down(token, cooldown_minutes=120):
-    """Check if an asset is in cooldown after a Phase 1 exit."""
-    cooldowns = load_cooldowns()
-    if token not in cooldowns:
-        return False
-    exit_ts = cooldowns[token].get("exitTimestamp", 0)
-    elapsed_min = (now_ts() - exit_ts) / 60
-    return elapsed_min < cooldown_minutes
-
-
-def set_asset_cooldown(token, reason="phase1_exit"):
-    """Set a cooldown on an asset after Phase 1 exit."""
-    cooldowns = load_cooldowns()
-    cooldowns[token] = {
-        "exitTimestamp": now_ts(),
-        "reason": reason,
-        "setAt": now_iso(),
-    }
-    save_cooldowns(cooldowns)
-
-
-# ─── Scanner History ─────────────────────────────────────────
-
-def load_scan_history():
+    Replaces v3.x mcporter subprocess. ~10-50× faster (~280ms vs
+    2.5-5s cold start). Same call signature as v3's mcporter_call so
+    legacy call sites (preserved from condor-scanner.py) keep working.
+    """
     try:
-        with open(HISTORY_FILE) as f:
-            return json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        return {"scans": []}
+        return _wrapper_client.mcp_call(tool, **params)
+    except Exception as e:  # noqa: BLE001
+        log_event("condor_mcp_call_failed", tool=tool, error=str(e))
+        return None
 
 
-def save_scan_history(history, max_scans=60):
-    if len(history["scans"]) > max_scans:
-        history["scans"] = history["scans"][-max_scans:]
-    atomic_write(HISTORY_FILE, history)
-
-
-# ─── MCP Helpers ─────────────────────────────────────────────
-
-def mcporter_call(tool, retries=2, timeout=25, **params):
-    """Call a Senpi MCP tool via mcporter."""
-    args = json.dumps(params) if params else "{}"
-    cmd = ["mcporter", "call", "senpi", tool, "--args", args]
-    for attempt in range(retries):
-        try:
-            r = subprocess.run(cmd, capture_output=True, text=True, timeout=timeout)
-            if r.returncode != 0:
-                if attempt < retries - 1:
-                    time.sleep(2)
-                    continue
-                return None
-            raw = json.loads(r.stdout)
-            if isinstance(raw, dict) and "content" in raw:
-                content = raw["content"]
-                if isinstance(content, list) and content:
-                    first = content[0]
-                    if isinstance(first, dict) and "text" in first:
-                        try:
-                            return json.loads(first["text"])
-                        except (json.JSONDecodeError, TypeError):
-                            pass
-            return raw
-        except subprocess.TimeoutExpired:
-            if attempt < retries - 1:
-                time.sleep(2)
-                continue
-            return None
-        except (json.JSONDecodeError, Exception):
-            return None
-    return None
+# Backward-compat alias — keeps any legacy `cfg.mcporter_call(...)` call
+# sites working even though the underlying transport is in-process now.
+mcporter_call = mcp_call
 
 
 def get_clearinghouse(wallet):
     if not wallet:
         return None
-    return mcporter_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
+    return mcp_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
 
 
 def get_positions(wallet):
+    """Returns (account_value, [position_dicts])."""
     ch = get_clearinghouse(wallet)
     if not ch:
         return 0, []
@@ -248,6 +196,10 @@ def get_positions(wallet):
 def output(data):
     print(json.dumps(data))
     sys.stdout.flush()
+
+
+def log(msg):
+    print(f"[condor-v4] {msg}", file=sys.stderr, flush=True)
 
 
 def now_ts():
